Save/Load fully implemented
Load finished
Added UpdateProgress method
Added Reset method
Slightly modified AddFiles method to increase intTotalFiles counter
Changed up the Directories ListView listener so that it first checks if
there was in fact, a change. This was done because if we tried using the
Reset method, we crashed since we tried to use an index that doesn't
exist.